Add render pass annotations (#8)
They show up in graphics debugger and group together the draw calls into their relevant pass, making it a lot easier to look through.

+++ Engine/Application.cpp
@@ -266,7 +266,7 @@ namespace Bat
 					}
 
 					static int selected_clip = 0;
-					ImGui::Combo( "Clip", &selected_clip, clip_names.data(), clip_names.size() );
+					ImGui::Combo( "Clip", &selected_clip, clip_names.data(), (int)clip_names.size() );
 					if( ImGui::Button( "Add clip" ) )
 					{
 						AnimationState new_state( clips[selected_clip] );
